Backmerged Dan's changes on hashcodes + generating IDs.  - This has an implication for the Axis WSDOAllSender, where the wssConfig must not be created each time a request is made.  - This is for the scenario that two handlers are chained together...they'll have the same ID's if wssConfig is created per handler.

+++ src/org/apache/ws/security/message/WSSecSignature.java
@@ -289,7 +289,7 @@ public class WSSecSignature extends WSSecBase {
         //
         crypto = cr;
         document = doc;
-        wsDocInfo = new WSDocInfo(doc.hashCode());
+        wsDocInfo = new WSDocInfo(doc);
         wsDocInfo.setCrypto(cr);
 
         //
@@ -314,7 +314,7 @@ public class WSSecSignature extends WSSecBase {
                     new Object[] { user, "signature" }
                 );
             }
-            certUri = "CertId-" + certs[0].hashCode();
+            certUri = wssConfig.getIdAllocator().createSecureId("CertId-", certs[0]);  
             //
             // If no signature algorithm was set try to detect it according to the
             // data stored in the certificate.
@@ -371,17 +371,16 @@ public class WSSecSignature extends WSSecBase {
         }
 
         sig.addResourceResolver(EnvelopeIdResolver.getInstance());
-        String sigUri = "Signature-" + sig.hashCode();
-        sig.setId(sigUri);
+        sig.setId(wssConfig.getIdAllocator().createId("Signature-", sig));
 
         keyInfo = sig.getKeyInfo();
-        keyInfoUri = "KeyId-" + keyInfo.hashCode();
+        keyInfoUri = wssConfig.getIdAllocator().createSecureId("KeyId-", keyInfo);
         keyInfo.setId(keyInfoUri);
 
         secRef = new SecurityTokenReference(doc);
-        strUri = "STRId-" + secRef.hashCode();
+        strUri = wssConfig.getIdAllocator().createSecureId("STRId-", secRef);
         secRef.setID(strUri);
-
+        
         //
         // Prepare and setup the token references for this Signature
         //
@@ -691,7 +690,7 @@ public class WSSecSignature extends WSSecBase {
      * @throws WSSecurityException
      */
     public void computeSignature() throws WSSecurityException {
-        WSDocInfoStore.store(wsDocInfo);
+        boolean remove = WSDocInfoStore.store(wsDocInfo);
         try {
             if (keyIdentifierType == WSConstants.UT_SIGNING ||
                     keyIdentifierType == WSConstants.CUSTOM_SYMM_SIGNING ||
@@ -716,7 +715,9 @@ public class WSSecSignature extends WSSecBase {
                 WSSecurityException.FAILED_SIGNATURE, null, null, e1
             );
         } finally {
-            WSDocInfoStore.delete(wsDocInfo);
+            if (remove) {
+                WSDocInfoStore.delete(wsDocInfo);
+            }
         }
 
     }
